// apps/erp/src/types/mfg.ts
export interface BomComponent {
  item_id: string;
  qty: number;
  uom: string;
  expected_scrap_pct: number;
  _edited?: boolean;
}